Pass 16: Clippy cleanup — workspace passes -D warnings with zero errors
Makes the entire workspace pass `cargo clippy --workspace -- -D warnings`.

loki-layout:
- Remove redundant `as f32` cast (value is already f32)

loki-odf:
- Remove unused `CellProps` import
- Collapse nested `if let` / `if` into `if let … && …` (3 sites)
- `map().unwrap_or()` → `map_or()` for ColWidth resolution
- `repeat().take()` → `repeat_n()` for column spec iteration
- Change `&Option<Vec<T>>` parameter to `Option<&Vec<T>>` + update 6 call sites
- Redundant closure `.filter_map(|ts| map_tab_stop(ts))` → `.filter_map(map_tab_stop)`
- `#[allow(clippy::type_complexity)]` on complex return-type function
- `Clone::clone(&x)` → `x.clone_from(&src)` for border shorthand assignments

loki-ooxml (mapper):
- `#[allow(clippy::too_many_arguments, clippy::too_many_lines)]` on map_document
- `#[allow(clippy::cast_precision_loss)]` for twip→f32 conversions (values small enough)
- i32→f64 casts replaced with `f64::from()` (4 sites, always lossless)
- Doc comment: wrap `row_span` in backticks
- `#[allow(clippy::type_complexity)]` on compute_v_merge_spans
- Redundant closure `|r| r.len()` → `Vec::len`

loki-ooxml (write/assembly):
- `#[allow(clippy::too_many_lines)]` on assemble_docx
- `!has_*` negated if-else → positive form (3 sites)
- Remove redundant `&` from `PartName::new(&format!(…))` (3 sites)

loki-ooxml (write/document):
- `#[allow(clippy::similar_names)]` on write_styled_para (has_pp/has_cp)
- `#[allow(clippy::too_many_lines, unused_assignments)]` on write_para_props_inline
- Use explicit `PageOrientation::Portrait` instead of wildcard
- Merge `BlockQuote` and `Div` identical match arms with `|`
- Remove `NotesBlock | RawBlock` explicit arm (covered by wildcard)
- Remove duplicate `ParagraphAlignment::Left` arm (covered by `_`)
- Remove duplicate `TabAlignment::Left` arm (covered by `_`)
- Remove duplicate `TabLeader::None` arm (covered by `_`)
- `Spacing::Percent(_) | _` pattern for #[non_exhaustive] Spacing enum
- `let mut line_s = String::new()` + empty `_ =>` to fix unused-assignment
- `#[allow(clippy::struct_excessive_bools)]` on RunProps
- `#[allow(clippy::too_many_lines)]` on write_inline
- Merge `Inline::Cite` and `Inline::Span` identical arms with `|`
- Remove duplicate `NoteKind::Footnote` arm from elem match
- Remove `Field | Comment | RawInline` explicit arm (covered by wildcard)
- Pass `BookmarkKind` by value (it's Copy); update call site with `*kind`
- `#[allow(clippy::unnecessary_wraps, clippy::similar_names)]` on write_inline_drawing
- Merge `Inline::Str(t)` with `Code(_, t) | Math(_, t)` into one arm
- `914_400` literal separators

loki-ooxml (write/numbering):
- `match` for single pattern → `if let … else { … }`
- `#[allow(clippy::cast_sign_loss)]` on `.max(1) as u32` (always ≥1)

loki-ooxml (write/styles):
- Remove duplicate `ParagraphAlignment::Left` arm (3 match sites)
- `Spacing::Percent(_) | _` for #[non_exhaustive] Spacing (2 sites)
- Collapse nested `if let` / `if let` into `if let … && let …`

loki-text:
- `#[allow(clippy::too_many_arguments)]` on home_tab_content

CLAUDE.md: Add Clippy compliance section to coding conventions.

+++ loki-layout/src/para.rs
@@ -594,7 +594,7 @@ pub fn layout_paragraph(
         let first_baseline = phantom
             .lines()
             .next()
-            .map(|l| l.metrics().baseline as f32)
+            .map(|l| l.metrics().baseline)
             .unwrap_or(0.0);
         return ParagraphLayout {
             height: 0.0,
